zebra: support for masterless VTEPs
Currently zebra only implements `advertise-all-vni` functionality for
VTEPs that are attached to a bridge. Vteps that do not belong to a
bridge are ignored. However, there are cases where it is useful for such
VTEPs to be included in `advertise-all-vni` functionality, such as
provisioning host services directly through VTEP devices that have IP
addresses.

This commit does 3 things
    1. Notifies EVPN clients of a new VNI when a masterless VTEP is
       added.
    2. Notifies EVPN clients of the MAC address of a masterless VTEPs.
    3. Removes logic that precludes BUM entries from being added for
       masterless VTEPs.

Signed-off-by: Ryan Goodfellow <rgoodfel@isi.edu>